refactor(moss-tts): consolidate audio tokenizer runtime
